Add Semantic Analysis interface and test analyzer
This function for now only covers some very specific parts of the
semantic analysis process. It will probably expand with time as it seems
a good idea to keep it all in the same class, when using a recursive
approach like this one. In the future it may be modified, expanded and
even renamed, if it goes out of its own scope